+++ js/gtUINodes.js
@@ -55,20 +55,20 @@ app.registerExtension({
     app.ui.settings.addSetting({
       id: `Griptape.allow_code_execution_dangerous`,
       category: ["Griptape", "!Griptape", "code_execution_dangerous"],
-      name: "Enable Insecure Code Execution [DANGER]",
+      name: "Enable Insecure Griptape Code: Run Python [DANGER]",
       type: "boolean",
-      tooltip: "When enabled, the Griptape Util: Code Execution node not check for dangerous code.\n\n[WARNING] This setting is dangerous and should only be enabled if you know what you are doing.",
+      tooltip: "When enabled, the Griptape Code: Run Python node will not check for dangerous code.\n\n[WARNING] This setting is dangerous and should only be enabled if you know what you are doing.",
       defaultValue: false,
-      onChange: (newVal, oldVal) => { if (newVal == true) { console.warn("Griptape Utils: Dangerous Code Execution enabled: ", newVal)} },
+      onChange: (newVal, oldVal) => { if (newVal == true) { console.warn("Griptape Code: Dangerous Code Execution enabled: ", newVal)} },
     });
     app.ui.settings.addSetting({
       id: `Griptape.allow_code_execution`,
       category: ["Griptape", "!Griptape", "code_execution"],
-      name: "Enable Code Execution Nodes",
+      name: "Enable Griptape Code: Run Python Nodes",
       type: "boolean",
-      tooltip: "When enabled, the Griptape Util: Code Execution node will be available for use.",
+      tooltip: "When enabled, the `Griptape Code: Run Python` node will be available for use.",
       defaultValue: false,
-      onChange: (newVal, oldVal) => { console.log("Setting got changed!", newVal) },
+      // onChange: (newVal, oldVal) => { console.log("Setting got changed!", newVal) },
     });
     Object.entries(keys_organized).forEach(([category, keys]) => {
       keys.forEach((key) => {